:truck: infrastructure 레이어 package 구조 변경
