--- src/main/java/com/hhplus/concertReserv/infrastructure/member/jpaRepository/MemberJPARepository.java
@@ -1,10 +1,0 @@
-package com.hhplus.concertReserv.infrastructure.member.jpaRepository;
-
-import com.hhplus.concertReserv.domain.member.entity.Users;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface MemberJPARepository extends JpaRepository<Users, UUID> {
-
-}
