+++ src/main/java/com/hhplus/concertReserv/infrastructure/redis/token/redisRepository/TokenRedisRepository.java
@@ -1,0 +1,7 @@
+package com.hhplus.concertReserv.infrastructure.redis.token.redisRepository;
+
+import com.hhplus.concertReserv.domain.token.entity.Token;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TokenRedisRepository extends CrudRepository<Token,Long> {
+}
