--- src/main/java/com/hhplus/concertReserv/infrastructure/reservation/jpaRepository/ReservationJPARepository.java
@@ -1,10 +1,0 @@
-package com.hhplus.concertReserv.infrastructure.reservation.jpaRepository;
-
-import com.hhplus.concertReserv.domain.reservation.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface ReservationJPARepository extends JpaRepository<Reservation, UUID> {
-
-}
