+++ src/main/java/com/hhplus/concertReserv/infrastructure/spring/reservation/jpaRepository/ReservationJPARepository.java
@@ -1,0 +1,10 @@
+package com.hhplus.concertReserv.infrastructure.spring.reservation.jpaRepository;
+
+import com.hhplus.concertReserv.domain.reservation.entity.Reservation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface ReservationJPARepository extends JpaRepository<Reservation, UUID> {
+
+}
